# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Blocker X - Discord Bot Hosting Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserPlan = typeof UserPlan[keyof typeof UserPlan];


export const UserPlan = {
  free: 'free',
  premium: 'premium',
} as const;

export interface User {
  id: string;
  discordId: string;
  username: string;
  discriminator?: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  email?: string | null;
  plan: UserPlan;
  isAdmin: boolean;
  isBanned: boolean;
  hasInvite: boolean;
  createdAt: string;
  /** @nullable */
  lastLogin?: string | null;
}

export interface UserList {
  users: User[];
  total: number;
}

export interface Profile {
  user: User;
  botCount: number;
  storageUsedBytes: number;
  aiUsageCount: number;
  /** @nullable */
  aiUsageLimit?: number | null;
  deploymentCount: number;
}

export interface ProfileUpdate {
  timezone?: string;
  notifications?: boolean;
}

export interface InviteCodeInput {
  code: string;
}

export interface InviteCode {
  id: string;
  code: string;
  /** @nullable */
  maxUses?: number | null;
  usesCount: number;
  /** @nullable */
  expiresAt?: string | null;
  isActive: boolean;
  createdBy: string;
  createdAt: string;
}

export interface InviteCodeCreate {
  /** @nullable */
  maxUses?: number | null;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  customCode?: string | null;
  grantsPremium?: boolean;
}

export type BotLanguage = typeof BotLanguage[keyof typeof BotLanguage];


export const BotLanguage = {
  python: 'python',
  javascript: 'javascript',
} as const;

export type BotStatusProperty = typeof BotStatusProperty[keyof typeof BotStatusProperty];


export const BotStatusProperty = {
  stopped: 'stopped',
  starting: 'starting',
  running: 'running',
  errored: 'errored',
  deploying: 'deploying',
} as const;

export interface Bot {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  language: BotLanguage;
  status: BotStatusProperty;
  userId: string;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  mainFile?: string | null;
  r2Prefix: string;
}

export type BotInputLanguage = typeof BotInputLanguage[keyof typeof BotInputLanguage];


export const BotInputLanguage = {
  python: 'python',
  javascript: 'javascript',
} as const;

export interface BotInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  language: BotInputLanguage;
  mainFile?: string;
  token: string;
  clientId: string;
  clientSecret: string;
}

export interface BotUpdate {
  name?: string;
  description?: string;
  mainFile?: string;
}

export interface BotStatus {
  botId: string;
  status: string;
  /** @nullable */
  uptime?: number | null;
  /** @nullable */
  memoryMB?: number | null;
  /** @nullable */
  lastStarted?: string | null;
  /** @nullable */
  lastStopped?: string | null;
}

export type DeploymentStatus = typeof DeploymentStatus[keyof typeof DeploymentStatus];


export const DeploymentStatus = {
  pending: 'pending',
  running: 'running',
  success: 'success',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;

export interface Deployment {
  id: string;
  botId: string;
  botName: string;
  userId: string;
  status: DeploymentStatus;
  startedAt: string;
  /** @nullable */
  finishedAt?: string | null;
  /** @nullable */
  logs?: string | null;
  /** @nullable */
  errorMessage?: string | null;
}

export type FileEntryType = typeof FileEntryType[keyof typeof FileEntryType];


export const FileEntryType = {
  file: 'file',
  directory: 'directory',
} as const;

export interface FileEntry {
  name: string;
  path: string;
  type: FileEntryType;
  /** @nullable */
  size?: number | null;
  /** @nullable */
  lastModified?: string | null;
}

export interface FilePathInput {
  path: string;
}

export interface FileRenameInput {
  oldPath: string;
  newPath: string;
}

export interface FileContent {
  path: string;
  content: string;
  encoding?: string;
}

export interface FileWriteInput {
  path: string;
  content: string;
}

export interface FileUploadInput {
  path: string;
  name: string;
  content: string;
  encoding?: string;
}

export interface EnvVar {
  id: string;
  botId: string;
  key: string;
  value: string;
  createdAt: string;
}

export interface EnvVarInput {
  key: string;
  value: string;
}

export type LogEntryLevel = typeof LogEntryLevel[keyof typeof LogEntryLevel];


export const LogEntryLevel = {
  info: 'info',
  warn: 'warn',
  error: 'error',
  debug: 'debug',
} as const;

export interface LogEntry {
  id: string;
  botId: string;
  level: LogEntryLevel;
  message: string;
  timestamp: string;
}

export type AIChatInputLanguage = typeof AIChatInputLanguage[keyof typeof AIChatInputLanguage];


export const AIChatInputLanguage = {
  python: 'python',
  javascript: 'javascript',
} as const;

export interface AIChatInput {
  message: string;
  /** @nullable */
  botId?: string | null;
  /** @nullable */
  filePath?: string | null;
  language?: AIChatInputLanguage;
  /** @nullable */
  context?: string | null;
}

export interface AIChatResponse {
  response: string;
  tokensUsed: number;
  usageCount: number;
  /** @nullable */
  usageLimit?: number | null;
}

export interface AIUsage {
  count: number;
  /** @nullable */
  limit?: number | null;
  plan: string;
}

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  info: 'info',
  success: 'success',
  warning: 'warning',
  error: 'error',
  announcement: 'announcement',
} as const;

export interface Notification {
  id: string;
  userId: string;
  title: string;
  message: string;
  type: NotificationType;
  isRead: boolean;
  createdAt: string;
}

export interface DashboardStats {
  totalBots: number;
  runningBots: number;
  totalDeployments: number;
  storageUsedBytes: number;
  aiUsageCount: number;
  recentDeployments: Deployment[];
  recentLogs: LogEntry[];
}

export type StorageStatsBotsItem = {
  botId: string;
  botName: string;
  usedBytes: number;
};

export interface StorageStats {
  usedBytes: number;
  limitBytes: number;
  fileCount: number;
  bots: StorageStatsBotsItem[];
}

export interface AdminStats {
  totalUsers: number;
  premiumUsers: number;
  bannedUsers: number;
  totalBots: number;
  runningBots: number;
  totalDeployments: number;
  activeInviteCodes: number;
  totalAIRequests: number;
  onlineUsers: number;
}

export type AnnouncementInputType = typeof AnnouncementInputType[keyof typeof AnnouncementInputType];


export const AnnouncementInputType = {
  info: 'info',
  success: 'success',
  warning: 'warning',
  error: 'error',
  announcement: 'announcement',
} as const;

export interface AnnouncementInput {
  title: string;
  message: string;
  type?: AnnouncementInputType;
}

export interface AuditLog {
  id: string;
  /** @nullable */
  userId?: string | null;
  action: string;
  /** @nullable */
  target?: string | null;
  /** @nullable */
  details?: string | null;
  /** @nullable */
  ip?: string | null;
  createdAt: string;
}

export type DiscordCallbackParams = {
code: string;
state?: string;
};

export type ListFilesParams = {
dirPath?: string;
};

export type UploadFileParams = {
path?: string;
};

export type ReadFileParams = {
filePath: string;
};

export type GetBotLogsParams = {
logLimit?: number;
};

export type ListUsersParams = {
search?: string;
limit?: number;
offset?: number;
};

export type ListInviteCodesParams = {
search?: string;
};

export type GetSystemLogsParams = {
logLimit?: number;
};