# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Blocker X - Discord Bot Hosting Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserPlan } from './userPlan';

export interface User {
  id: string;
  discordId: string;
  username: string;
  discriminator?: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  email?: string | null;
  plan: UserPlan;
  isAdmin: boolean;
  isBanned: boolean;
  hasInvite: boolean;
  createdAt: string;
  /** @nullable */
  lastLogin?: string | null;
}